Moves and adds

+++ internal.h
@@ -81,7 +81,7 @@ struct rcksum_state {
     } stats;
 
 	list<size_t> *offsets;
-	map<size_t, list<size_t> > *moves;
+	map<long long, list<size_t> > *moves;
 	map<size_t, size_t> *add;
 	map<size_t, size_t> *del;
 };
